+++ server/package.json
@@ -20,7 +20,9 @@
   },
   "homepage": "https://github.com/DYGITArchitect/backend#readme",
   "dependencies": {
+    "cookie-parser": "^1.4.6",
     "cors": "^2.8.5",
+    "dotenv": "^16.0.3",
     "express": "^4.18.2",
     "express-fileupload": "^1.4.0",
     "mongoose": "^6.7.2",
