+++ server/routers/router.js
@@ -1,0 +1,12 @@
+import { Router } from "express";
+import PostController from "../controllers/PostController.js";
+
+const router = new Router();
+
+router.post("/posts", PostController.create);
+router.get("/posts", PostController.getAll);
+router.get("/posts/:id", PostController.getOne);
+router.put("/posts", PostController.update);
+router.delete("/posts/:id", PostController.delete);
+
+export default router;
