+++ server/services/PostService.js
@@ -1,5 +1,5 @@
-import Post from "./Post.js";
-import fileService from "./fileService.js";
+import Post from "../models/Post.js";
+import fileService from "../services/fileService.js";
 
 class PostService {
   async create(post, picture) {
